+++ RTS Project code/Motor2D/Render.cpp
@@ -57,8 +57,8 @@ bool Render::Awake(pugi::xml_node& config)									//Renderer VSync --> Flags tu
 	{
 		camera.w = App->win->screen_surface->w;
 		camera.h = App->win->screen_surface->h;
-		camera.x = config.child("camera").attribute("x").as_float();
-		camera.y = config.child("camera").attribute("y").as_float();
+		camera.x = (int)config.child("camera").attribute("x").as_float();
+		camera.y = (int)config.child("camera").attribute("y").as_float();
 
 		//cam.camera_debug = config.child("camera_debug").attribute("value").as_bool();				//Loads the cameraLimit bool state from the config xml file.
 		//cam.camera_debug_speed = config.child("camera_debug").attribute("speed").as_float();		//Loads the camera's movement speed value while in debug mode  from the config xml file.
@@ -149,7 +149,7 @@ void Render::ResetViewPort()
 iPoint Render::ScreenToWorld(int x, int y) const
 {
 	iPoint ret;
-	int scale = App->win->GetScale();
+	int scale = (int)App->win->GetScale();
 
 	ret.x = (x - camera.x / scale);
 	ret.y = (y - camera.y / scale);
@@ -165,8 +165,8 @@ bool Render::Blit(SDL_Texture* texture, int x, int y, const SDL_Rect* section, b
 
 	SDL_Rect rect;
 
-	rect.x = (int)(camera.x * speed) + x * scale;
-	rect.y = (int)(camera.y * speed) + y * scale;
+	rect.x = (int)(camera.x * speed) + x * (int)scale;
+	rect.y = (int)(camera.y * speed) + y * (int)scale;
 
 	if (section != NULL)
 	{
@@ -178,8 +178,8 @@ bool Render::Blit(SDL_Texture* texture, int x, int y, const SDL_Rect* section, b
 		SDL_QueryTexture(texture, NULL, NULL, &rect.w, &rect.h);
 	}
 
-	rect.w *= scale * render_scale;
-	rect.h *= scale * render_scale;
+	rect.w *= (int)(scale * render_scale);
+	rect.h *= (int)(scale * render_scale);
 
 
 	SDL_Point* p = NULL;
@@ -204,7 +204,7 @@ bool Render::Blit(SDL_Texture* texture, int x, int y, const SDL_Rect* section, b
 bool Render::DrawQuad(const SDL_Rect& collider, Uint8 r, Uint8 g, Uint8 b, Uint8 a, bool filled, bool use_camera) const
 {
 	bool ret = true;
-	uint scale = App->win->GetScale();
+	uint scale = (int)App->win->GetScale();
 
 	SDL_SetRenderDrawBlendMode(renderer, SDL_BLENDMODE_BLEND);
 	SDL_SetRenderDrawColor(renderer, r, g, b, a);
@@ -232,7 +232,7 @@ bool Render::DrawQuad(const SDL_Rect& collider, Uint8 r, Uint8 g, Uint8 b, Uint8
 bool Render::DrawLine(int x1, int y1, int x2, int y2, Uint8 r, Uint8 g, Uint8 b, Uint8 a, bool use_camera) const
 {
 	bool ret = true;
-	uint scale = App->win->GetScale();
+	uint scale = (int)App->win->GetScale();
 
 	SDL_SetRenderDrawBlendMode(renderer, SDL_BLENDMODE_BLEND);
 	SDL_SetRenderDrawColor(renderer, r, g, b, a);
@@ -256,7 +256,7 @@ bool Render::DrawLine(int x1, int y1, int x2, int y2, Uint8 r, Uint8 g, Uint8 b,
 bool Render::DrawCircle(int x, int y, int radius, Uint8 r, Uint8 g, Uint8 b, Uint8 a, bool use_camera) const
 {
 	bool ret = true;
-	uint scale = App->win->GetScale();
+	uint scale = (int)App->win->GetScale();
 
 	SDL_SetRenderDrawBlendMode(renderer, SDL_BLENDMODE_BLEND);
 	SDL_SetRenderDrawColor(renderer, r, g, b, a);
